+++ wai-extra/Network/Wai/Middleware/Routed.hs
@@ -1,4 +1,10 @@
-module Network.Wai.Middleware.Routed where
+-- |
+--
+-- Since 3.0.9
+module Network.Wai.Middleware.Routed
+    ( routedMiddleware
+    , hostedMiddleware
+    ) where
 
 import Network.Wai
 import Data.ByteString (ByteString)
@@ -10,6 +16,7 @@ import Data.Text (Text)
 --
 -- > let corsify = routedMiddleWare ("static" `elem`) addCorsHeaders
 --
+-- Since 3.0.9
 routedMiddleware :: ([Text] -> Bool) -- ^ Only use middleware if this pathInfo test returns True
                  -> Middleware -- ^ middleware to apply the path prefix guard to
                  -> Middleware -- ^ modified middleware
@@ -18,6 +25,8 @@ routedMiddleware pathCheck middle app req
   | otherwise                = app req
 
 -- | Only apply the middleware to certain hosts
+--
+-- Since 3.0.9
 hostedMiddleware :: ByteString -- ^ Domain the middleware applies to
                  -> Middleware -- ^ middleware to apply the path prefix guard to
                  -> Middleware -- ^ modified middleware